refactor(output): extract group tests to sibling file
